Make nrvo-string test more robust.
The breakpoint locations were in places where clang doesn't actually
emit a source location for and depend on the debugger's ability to
move the breakpoint forward onto a line that is already in the
function epilogue. In my testing older versions of LLDB fail to do
that, so I'm modifying the test to force a break-able location by
calling a noinline function.

<rdar://problem/52079841>

git-svn-id: https://llvm.org/svn/llvm-project/debuginfo-tests/trunk@364466 91177308-0d34-0410-b5e6-96231b3b80d8